Fix CI jobs to actually work with PRs from forks (#79)
In #58 new CI jobs were added to attempt to build and publish a live
version of the docs from pull requests and merged commits with github
pages. However, this had 2 issues, the first is that the auth model in
github doesn't allow PRs from forks write access (which is a security
precaution). The second is that multiple PRs will interfere with each
other and overwrite each other if the jobs are triggered at the same
time.

This commit fixes this issue by changing the CI configuration to only
build the docs in CI and store the built documentation as a zip file
artifact that can be downloaded. While the actualy "live version"
publish job only runs post-merge.