Remove incrementors from language specification
This removes the increment-and-return operators '++' and '--' from the
reference text and grammar.  The text did not specify whether these
should be pre- or post-incrementors, though only post-incrementors were
allowed by the ANTLR grammar.

The TSC conensus is that while these operators can easily be defined
unambiguously by parsers, in practice they are less readable by humans
than the equivalent `x += 1;` statement, and there is no significant
benefit to having them in OpenQASM 3.

See #144.