Refactor structure of tests
This completely refactors the `test_grammar.py` file, to use the
tree-printing function from the reference parser package and to
recursively discover the test files, rather than having them hardcoded.
This latter point makes it easier to add tests; now, putting a YAML file
in the `reference` directory will guarantee that it gets run, and
doesn't rely on remembering to add it somewhere in the testing file as
well.