Refactor invalid statement cases into files
Having the cases be specified directly in parametrised lists inside
pytest makes it harder to reuse the tests for other purposes.  Instead,
we separate them out into grouped files, with individual cases on
separate lines.

The parametrisation of the test cases is set up to report the file name
and line number in the id for easier tracking down of the relevant
location, since pytest will automatically report the string value of the
statement in its output on a test failure.

We keep several statements in one file to make it easier to read all the
tests at a glance; it's annoying and time-consuming to open many
different files to see the cases, and that also means you have to name
each file.  Also, having many different files means the testing process
would have to thrash the disk IO, which really isn't ideal for simple
cases.