Add packaging structure to reference AST
This creates a full Python package (called `openqasm3`) out of the
reference AST, parser and visitor, so it can be installed through `pip`
in the normal manner, and distributed through PyPI.

The `test_openqasm_tests.py` file is not currently included when
generating the sdist, because it relies on files that are elsewhere in
the repository, and can't be brought into this package without
significant duplication.  This may be changed in the future, but for
now, the test file is useless on its own.

The minimum Python version is fixed at Python 3.7; to support Python
3.6, there would need to be a (3.6-specific) dependency on a backport of
`dataclasses`, but also the `from __future__ import annotations` does
not exist, so the entire `ast.py` file would need re-ordering into a
bottom-up approach, which harms legibility for little gain.