Swap concatenation operator to ++ from ||
This is the consensus of the TSC from 2021-09-16, in order to avoid a
clash between concatenation and logical or.  This operator is similar to
that used in Haskell and Scala for the same purpose.  Originally,
OpenQASM 3 had the pre- and post-increment operator which would also
have clashed with ++, but this is no longer the case, so the token is
free for clash-free usage.