Add multi-dimensional arrays to reference grammar
This necessitated more wide-ranging changes than might have been
expected, because it needed to overcome various prior assumptions that
arrays would be single-dimensioned only.  There were portions where
`indexIdentifier` was defined in order to be used in `let` aliases, but
then reused in various other places.  This caused issues where things
like `a ++ b` were valid in places where it should not have been, and
regular indexing (`a[2]`) wasn't valid in places where it should have
been (`5[4]`, for example).

Aside from the obvious changes of adding `array` and the necessary
machinery to make this work in declarations and subroutine definitions,
this commit also separates out `indexedIdentifier` into its own entry.
This makes arbitrary levels of nested indexing on identifiers allowed in
any place that an expression is allowed.  Rules which needed to take a
simple identifier (but not an expression), such as quantum reset
operations are changed to use the new `indexedIdentifier` instead of the
old alias-based version.  This makes `reset a ++ b;` into an error,
whereas previously it was allowed by the grammar; there is currently no
basis in the specification to allow this (though it would be a simple
addition).

The separation of the `indexedIdentifier` was also necessary to make
assignment into array slices work correctly.  Before, the `designator`
rule was overloaded and used in assignments, despite it being meant for
a completely different part of the language.  An `indexedIdentifier` is
now valid as an l-value for assignments, which allows sliced,
multiple-dimensioned assignments such as

    my_array[0, 1:3] = my_array[1, 4:6];

There are some places where `indexedIdentifier` is used, though it is
possible that `indexExpression` may be more appropriate, such as the
operands of quantum gates.  For now, these are left as-is to reduce the
changeset.

Co-authored-by: Michael Healy <mbhealy@ibm.com>