Require braces in discrete-set indexing
This changes the syntax of indexing a discrete set out of a register, by
requiring explicit braces around the list.  For example, what was
previously

    let alias = register[1, 2, 3];

is now

    let alias = register[{1, 2, 3}];

The reason for this change is to avoid a confusing syntax conflict with
multi-dimensional arrays, as they will use the old form to be
multi-dimensional indexing.  While arrays and aliased registers may well
be distinct in the type system, it would still be confusing for
programmers to have the same syntax mean two quite separate things.