Clarify meaning of `angle` type
This expands the spec with more detailed descriptions of the effective
bit representations of the `angle` type, and expands on the reasoning
for the introduction of the type, and how various mathematical
operations apply on it.  This is meant as a clarification of the
specification; it is attempting to describe the same behaviour as
before, but in more precise terms.

It is worth mentioning that hardware is not _compelled_ to implement the
angle type in a hardware register that stores these exact bit patterns
in memory.  It is only required that the logical interactions with the
``bit`` and other types, and bit-indexing operations return the same
values as described by the specification.