Update grammar for typed for-loop variables
This updates the ANTLR grammar to require an explicit type for for-loop
variables.  This is a breaking change from the previous language
specification, so it's accompanied by changes in the example files.

The reference AST and vim plugins are also updated.  The reference AST
visitor learned a new option `permissive` in its `parse` function, since
it wasn't actually failing its tests when the ANTLR grammar required the
type of the variable, but the AST had not yet been updated.  This was
because ANTLR attempts to continue through a missing token and only
emits a warning message.  Turning this on also turned up a missing
semicolon in the `durationof` test.