Split lexer and parser ANTLR grammars
This makes a clean split of the lexer and parser rules for the ANTLR
grammar, with no changes to the actual grammar of the language other
than those that were required to make the split.  The small extent of
these changes can be seen in the small diffs of the reference outputs.
The purpose of this split was to enable us to use lexer-specific modes
in the reference implementation, such as island grammars for parsing
pragmas and `defcal` blocks.

The main changes are just that keywords, symbols and builtin names are
now lexed as explicit tokens, rather than having implicit rules within
the parser.  In a couple of cases, moving the rules from unnecessary
parser rules (e.g.  booleanLiteral) into lexer rules caused a small part
of the parse tree to collapse, for example

    booleanLiteral
        true

became

    true

since the parser's `booleanLiteral` became a lexeme `BooleanLiteral`.

This should not have made any previously valid QASM 3 programmes
invalid, nor the other way around.