Correct type of arguments in QuantumGateDefinition
The ANTLR to AST conversion was taking the type of the arguments to be
`Identifier`, but the AST hierarchy itself wanted them to be
`ClassicalArgument`.  The spec currently says that they are
"arbitrary-precision numerical types", which does not line up with
`ClassicalArgument` (which would include, e.g. `IntType`), and
`Identifier` is more correct for the current spec, though this may
change in the future.