Refactor reference ANTLR grammar and Python parser
This is a near-complete re-write of the reference ANTLR grammar (and
consequently the AST-generating Python parser) to vastly simplify the
logical structure of the reference grammar.  The aims of this refactor
are:

1. to make the ANTLR grammar easier for users to read

2. to make the structure of OpenQASM 3 programs clearer in the grammar,
   when used as a reference

3. unify handling of various rules that will be important for work
   from the Pragmas and Annotations working group

In particular, the first two goals push more work onto the reference AST
parser (such as refusing 'measure' statements inside gates), in favour
of letting the ANTLR grammar be clearer about defining "statements" and
"expressions".  This seemed like a fair trade, since the ANTLR parser
alone was never complete already.

In its previous incarnation, when it was attempting to be a stand-alone
tool, the raw ANTLR parser comprised its usefulness as a reference
(which is its primary goal) in order to attempt to forbid some invalid
semantics.  Hopefully, this new form will be more useful by letting
ANTLR's very readable grammars be a quick reference for program
_structure_, and the reference Python parser can provide more detail on
what is semantically forbidden.

There are some relatively minor changes to the reference AST.  In
several cases, these were simply removing restrictions that had
arbitrarily been placed on programs by the previous ANTLR grammar.
For example, `input`, `output` and `include` declarations no longer need
to be the first statements of the file.

In other cases, the AST was slimmed down by removing redundant structure
that stemmed from the previous ANTLR grammar, such as
`QuantumForInLoop`.  This was previously not providing much structure,
since it was not itself a `QuantumStatment`.  Some statements, such as
`reset` were marked as `QuantumInstruction`, which falsely implied that
they could be used inside gates; `reset` cannot occur inside a gate,
because it is non-unitary.