Explicitly disambiguate no-qubit gate calls from functions
`gphase` is a zero-qubit "gate" in the grammar, which allows an
ambiguity with expression statements that contain only a single function
call.  For example,

        gphase(pi);

and
        fn(pi);

are syntactically nearly identical.  The ANTLR grammar before this
commit was already correctly resolving the ambiguity in favour of the
`expressionStatement` (except if the call was `gphase`), because
`expressionStatement` is tried before `gateCallStatement` in the
`statement` definition.  `gphase` was still resolved correctly because
it was tokenised as a keyword and not an identifier, so could not match
the function-call rule.  This was part of the reason for defining the
`gphase` token separately, but was still a bit too magical to rely on.

The new form explicitly disambiguates the two cases, exploiting the fact
that `gphase` is the _only_ gate that can have zero arguments (and even
then, only if there is no `ctrl` or `negctrl` modifiers).