Restrict complex types to floating-point
This more completely defines the semantics for working with complex
types in OpenQASM 3.0, including restricting them to only complex
values.  We would have needed special-case rules to define things like
division between `complex[int]` types, and the hardware support and
targetted use cases did not seem to be strong enough to spend time doing
so for the first version of the language.  The promotion rules may also
have been unusual.