Add basic formatter to reference AST
This adds a basic print visitor to the reference AST, which can be used
as a (very simple) code formatter.  It outputs valid OpenQASM 3.0, as
opposed to the AST tree format that is automatically provided by the
`dataclasses` structure (e.g. call `repr` on a node).

The structure of the printer is very similar to the print visitor in
Qiskit Terra (which I also wrote), which is a simple single-pass printer
into the stream.  This is perhaps the simplest way of writing the
printer, but it doesn't leave much room for extensibility on look-ahead
(e.g. breaking lines in good places or dynamically choosing the spacing
around binary operators based on the precedence of pairs of operands).
It could be extended to handle these, but a general-purpose,
configurable formatter might be a better fit in a stand-alone project.