Remove vim plugin from mono repository (#360)
Now that we have the openqasm GitHub repository, we can move the vim
plugin (and any potential future plugins) into its own repository.  This
now exists as openqasm/openqasm.vim, and the complete git history of the
`plugins/vim` subdirectory is extracted into that repository (including
re-linking the issues numbers to the correct location).