Expand on all scoping definitions
This adds more text and formalism on what variables are in scope, and
expands on the types of scopes.  In particular, we identify and talk
about the differences between the scoping of subroutine/gate bodies and
the scoping of local blocks with regard to symbol visibility.

Most of this commit is just additional details and examples on top of
the previous commit in the PR.  There are some changes to pre-existing
text:

- Removed mention of the for-loop iteration variable being immutable -
  this behaviour was removed from the PR about for-loop variables.

- Changed the visibility of symbols within `def` and `gate` statements
  such that *only* `const` variables, subroutines/gates and built-ins
  are visible.  The reason to remove making mutable run-time values
  visible is to forbid closing over these variables, which may be more
  difficult for immediate implementations with no current motivating
  use-case.