Formalise rules for compile-time `const` expressions
These updates are generally expected to be consistent with what
implementations already do, but formalised completely into the language
specification.

This also formalises the idea of the expressions that are guaranteed to
be evaluated at compile-time, and thus which expressions are guaranteed
to be valid to use as widths in type declarations, and in qubit
dereferencing.  This change does not attempt to solve the issue of the
allowed "dynamic dereferencing" of qubits, but fully defines the
problem: a reference to a qubit is dynamic if the type of the indexing
value is not `const`.  Implementations of OpenQASM 3.0 are permitted to
treat this as an error.

This change also fleshes out the section on built-in "scientific
calculator" functions on constant expressions, adding specific rules for
how the particular overload is chosen.