Drop dollar sign from pulse wildcard identifier
This changes the wildcard identifier pattern from `$<identifier>` to
just being the identifier (for example `$q` to `q`).  Precise hardware
qubits such as `$0` are not affected.  The extra dollar sign did not
carry any information, since the wildcard applied to _all_ qubits, and
caused a strange discrepancy between what was a valid identifier in
different situations.