Correct pulse-grammar parsing in reference ANTLR
This teaches the ANTLR grammar to recognise `cal` statements (which were
previously entirely absent), and to allow symbolic expressions in place
of classical argument definitions in the argument lists of `defcal`
statements (as stated in the spec).

More completely, the tokenisation of blocks in the calibration updated
has been completely rewritten so that ANTLR tokenises everything between
a balanced `{}` into one single token.  The new requirement that the
brace block is balanced fixes a family of bugs where ANTLR would extend
the (supposedly non-greedy) `LBRACE .* RBRACE` match to "eat" parser
errors following a `defcal` block.