Correct surrounding OQ 3 code in OpenPulse examples (#374)
In some OpenPulse (or general pulse-grammar) examples where the embedded
of the calibration language inside OpenQASM 3 was being demonstrated,
the surrounding OQ 3 code was invalid.  This was mostly due to OpenPulse
types being given in `extern` statements; this would be an OpenPulse
extension, and so needs to be in `cal` block.

Examples which were _entirely_ in the pulse language are unchanged;
there's no advantage in wrapping everything in a `cal` statement, and it
just adds more noise for readers.