Fix pulse example QASM files (#376)
Some files that included `defcal` statements had not been updated to
keep track with changes to OpenQASM.  They were not caught in CI because
the only tests of them are that the reference ANTLR parser can
successfully parse them, and its current rule for handling `defcal`
blocks allows "true" syntax errors to be masked.

The files `defcal.qasm` and `t1.qasm` have simple syntactic errors that
are easy to update.  The file `pong.qasm` is very old, and based on a
version of the timing information constraints that long predates even
the original arXiv paper.  It cannot be reasonably brought up-to-date
without being re-written completely.  It's currently a discussion topic
as to whether the timing / stretch-resolution behaviour it is
exemplifying should be a part of the OQ 3.0 specification, so we remove
the file for now to unblock CI for correct OpenPulse implementations,
and defer to those discussions on replacing it in the future with a new
example file.