Clarify types and meanings of bit rotations
The first argument is the value whose bits are to be rotated, and the
second is the "distance" that the bits are moved.  The only allowable
types for `target` are bit arrays and unsigned integers (and any values
that implicitly cast to them).  The allowable types for `distance` are
anything that can be implicitly promoted to a signed-integer type
(`int[_]` and `uint[_]`, e.g.).