Add deployment pipeline for reference AST
This adds a GitHub Action for handling the build and deployment of the
reference Python package.  This should remove the possibility for user
error when building the package (especially important now that there are
many ANTLR versions to be built), and allows anybody with tag-push
access to the repository to cut a release.

The action verifies that the package builds, installs and imports, but
we leave testing to the CI that should have run before any PR was
merged.  We also verify that the tag version matches the version
information in the wheel; this is mostly just a sanity check that what
we're releasing is what we expect.  PyPI will reject any uploads for a
version that has already been released.