Handle arbitrary-string tokenisation by ANTLR lexer
This has the advantage over the previous approach (specifying extra
possible continuations in the parser) that the structural reference of
the parser remains clearer, and consumers of the ANTLR-generated parse
tree such as the `openqasm3` package do not need to add special handling
for the alternate `BitstringLiteral` output.

This also automatically updates the handling around `defcalgrammar` to
work with arbitrary strings as well.