Monkeypatch docutils to fix bibliography output
This applies the patch discussed in gh-455 dynamically on import of
`docutils`, in order to fix the broken HTML output from the
`..  bibliography` directive. We constrain `docutils` during our build
process to fix it to a version that is known-good with this change
applied.  The constraint may be dropped when docutils updates, as the
relevant patch is merged to trunk.