Fail workflow on documentation build failures
The documentation-build script did not propagate failures from
individual lines, which meant that a failure in the script could still
result in an attempted empty deployment, which deletes the whole website.
This ensures that the bash script propagates the failure through to the
workflow status, so the deployment does not trigger.

We also fix the failure that took out the website last time - a missing
transitive dependency (which isn't our fault, but is easy to work
around).