Update Python versions and dependencies in AST GitHub Actions
This normalise the AST action to only run the style and lint checks
once, on the newest Python version, so that differences between the
linters and formatters that might arise from bugfixes not backported to
the last supported package version for the oldest Python do not cause
spurious failures between the two branches.  This notably happened with
`black`, where the latest version compatible with Python 3.7 did not
contain a bugfix that later versions did have access to.

Further, this updates the latest supported version of Python to 3.12,
which matches reality, since previously we had left it at 3.10.