Document content of the standard library
This describes the effects of the `stdgates.inc` include file as part of
the language specification.  I deliberately did not include the
`examples/stdgates.inc` version of it in the specification to avoid
implications that it *must* be implemented with that manner, but
attempted to explain how implementations must behave *as if* if were
defined like this, so that implementations are free to interpret the
file in terms of `defcal` statements or the like (if they were to so
choose).

This commit also includes a brief Sphinx extension, to make it possible
to document OpenQASM gates in the full Sphinx structure.