Declare `bit`/`bool` equivalence and distinction from `bit[1]`
This formalises the idea that `bit[n]` is a sequence-like type of which
`bit[1]` is just the length-one variant.  This is distinct from the type
`bit`, which is the scalar of it.  The `bit[n]` type is expected to be
stored bit-packed (certainly when used in expression positions), which
is reflected in how it casts to other integer types.

Taken alone, it is fairly meaningless to attempt to have a "bit-aligned"
scalar type `bit`.  We don't attempt to define semantics for this in any
sort ABI for OpenQASM 3; the only places this would be observable from
within the program are in `array` and in `extern` calls, both of which
it is forbidden from appearing in.  Instead, the semantics of scalar bit
are simply set to have `bit` and `bool` be interchangeable in r-value
usage.

An additional line about `qubit[n]` being equivalent to `array[qubit,
n]` is deleted here, since some of the motivation for the rest of this
commit is considering the alignment concerns of the `bit` and `bool`
types; it doesn't really make sense to talk about classical alignment of
a `qubit` type, and there doesn't appear to be any need to allow
`array[qubit, n]` when `qubit[n]` already suffices.