Modify parser such that tests pass
I modified the parser so that the tests that I added would pass.
I did this in the way most obvious to me.
I am very much a beginner with ANTLR, so this is likely shoddy work,
but I wanted nonetheless to make it public for review while I try
to figure out how to make it better.