feat: Add error listener for lexer on parse (#576)
* feat: Add error listener for lexer on parse

When parsing, callers have the option to set a `permissive` flag to let
ANTLR attempt to recover from incorrect input. When this is unset, then
the parser error handler is set to the bail error strategy. This commit
adds a corresponding error listener for the lexer when it encounters
an error, such as an unrecognised token.

* docs: Add release note

* fix: Move release notes to correct location

* fix: format with black