Restrict base type of array to documented types
Previously, `bit` and `stretch` were allowed as the base type of the
array type. These two types are explicitly disallowed by the specification.

This commit modifies the parser to reject these invalid base types.
In the parser definition file, the base type of the array type is
changed from scalarType to a new rule arrayBaseType. The latter is just
a copy of scalarType with the invalid base types removed.

The type `bit` appeared in some parser tests as the array base type.
These tests have been removed.

Two tests have been added to verify that the parser to produces an error
if either `bit` or `stretch` appears as a base type.